backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0", # Added fastembed dependency
    "fastembed>=0.2.0",
]
